fix: just include SDL.h and avoid error on macOS
On macOS filesystem is file insensitive so the include of the local file
"sdl/sdl.h" from the project was conflicting with include <SDL/SDL.h> from
the SDL library.

To fix this problem and simplify we just include <SDL.h>. The pkg-config
returns the directory SDL where SDL.h is located so that seems to be the
right way to include SDL main header file.